refactor: Fix DI for Emission services
